+++ build.gradle
@@ -51,7 +51,7 @@ repositories {
     // mavenLocal()
 }
 
-mainClassName = 'tw.edu.ntu.lowerbound10hours.jlask.Main'
+mainClassName = 'tw.edu.ntu.lowerbound10hours.example.Main'
 
 defaultTasks 'run'
